models/erase-lama-big/requirements.txt
numpy<2.0
onnxruntime
Pillow
